moved LoginDTO to auth package and updated imports, removed unused JsonProperty annotations from AuthType
