+++ java-server/src/main/java/de/pse/oys/service/auth/AuthService.java
@@ -1,11 +1,10 @@
 package de.pse.oys.service.auth;
 
 import de.pse.oys.dto.auth.AuthResponseDTO;
-import de.pse.oys.dto.LoginDTO;
+import de.pse.oys.dto.auth.LoginDTO;
 import de.pse.oys.dto.RefreshTokenDTO;
 import de.pse.oys.persistence.UserRepository;
 
-import org.springframework.beans.factory.annotation.Value;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
